export styled cf-component-heading by default
