+++ packages/cf-component-heading/README.md
@@ -17,7 +17,9 @@ import { Heading, HeadingCaption } from 'cf-component-heading';
 const HeadingComponent = () => (
   <Heading size={2}>
     Look at this nice heading!
-    <HeadingCaption>It even has a nice HeadingCaption</HeadingCaption>
+    <HeadingCaption>
+      It even has a nice HeadingCaption
+    </HeadingCaption>
   </Heading>
 );
