+++ packages/cf-component-heading/example/basic/component.js
@@ -1,22 +1,13 @@
 import React from 'react';
-import {
-  Heading,
-  HeadingCaption,
-  HeadingTheme,
-  HeadingCaptionTheme
-} from 'cf-component-heading';
-import { applyTheme } from 'cf-style-container';
-
-const StyledHeading = applyTheme(Heading, HeadingTheme);
-const StyledHeadingCaption = applyTheme(HeadingCaption, HeadingCaptionTheme);
+import { Heading, HeadingCaption } from 'cf-component-heading';
 
 const HeadingComponent = () => (
-  <StyledHeading size={2}>
+  <Heading size={2}>
     Look at this nice heading!
-    <StyledHeadingCaption>
+    <HeadingCaption>
       It even has a nice HeadingCaption
-    </StyledHeadingCaption>
-  </StyledHeading>
+    </HeadingCaption>
+  </Heading>
 );
 
 export default HeadingComponent;
