+++ packages/cf-component-heading/src/index.js
@@ -1,6 +1,18 @@
-import Heading from './Heading';
-import HeadingCaption from './HeadingCaption';
+import HeadingUnstyled from './Heading';
+import HeadingCaptionUnstyled from './HeadingCaption';
 import HeadingTheme from './HeadingTheme';
 import HeadingCaptionTheme from './HeadingCaptionTheme';
 
-export { Heading, HeadingCaption, HeadingTheme, HeadingCaptionTheme };
+import { applyTheme } from 'cf-style-container';
+
+const Heading = applyTheme(HeadingUnstyled, HeadingTheme);
+const HeadingCaption = applyTheme(HeadingCaptionUnstyled, HeadingCaptionTheme);
+
+export {
+  Heading,
+  HeadingUnstyled,
+  HeadingCaption,
+  HeadingCaptionUnstyled,
+  HeadingTheme,
+  HeadingCaptionTheme
+};
