+++ packages/cf-component-heading/test/Heading.js
@@ -1,13 +1,7 @@
 import React from 'react';
-import {
-  Heading as HeadingUnstyled,
-  HeadingTheme
-} from '../../cf-component-heading/src/index';
+import { Heading } from '../../cf-component-heading/src/index';
 import renderer from 'react-test-renderer';
 import felaTestContext from '../../../felaTestContext';
-import { applyTheme } from 'cf-style-container';
-
-const Heading = applyTheme(HeadingUnstyled, HeadingTheme);
 
 test('should render size', () => {
   const component = renderer.create(
